refactor: move api folder under src/lib
