+++ src/pages/api/auth/[...nextauth].ts
@@ -1,7 +1,7 @@
 import NextAuth, { NextAuthOptions } from 'next-auth';
 import CredentialsProvider from "next-auth/providers/credentials";
 
-import { login } from '@/api';
+import { login } from '@/lib/api';
 
 export const authOptions: NextAuthOptions =  {
   providers: [
